Login / Logout buttons refactored
