+++ src/components/login/LoginButton.tsx
@@ -1,36 +1,44 @@
-import { CredentialResponse, GoogleLogin } from '@react-oauth/google';
-import { getMatchups } from '../../apiCalls/getMatchups';
-import { useAppDispatch } from '../../hooks/hooks';
-import { loginAction } from '../../actions/userActions';
+import * as React from "react";
+import Button from "@mui/material/Button";
+import Modal from "@mui/material/Modal";
+import Box from "@mui/material/Box";
+import LoginPage from "./LoginPage";
 
-export default function LoginButton({closeModal}: any) {
+const style = {
+	position: "absolute" as "absolute",
+	top: "50%",
+	left: "50%",
+	transform: "translate(-50%, -50%)",
+	width: 360,
+	bgcolor: "background.paper",
+	border: "2px solid #000",
+	boxShadow: 24,
+	p: 4,
+};
 
-    const dispatch = useAppDispatch();
+const LoginButton = () => {
+    const [open, setOpen] = React.useState(false);
+    const handleOpen = () => setOpen(true);
+    const handleClose = () => setOpen(false);
 
-    const onSuccess = (res: CredentialResponse) => {
-        console.log("Login Successful! Current user: " + JSON.stringify(res))
-        localStorage.setItem("user", JSON.stringify(res))
+    return (
+        <>
+            <Button color="inherit" onClick={handleOpen}>
+                Login
+            </Button>
 
-        dispatch(loginAction(res));
-        dispatch(getMatchups(res));
-        
-        closeModal();
-    }
-    const onError = (res: any) => {
-        console.log("Login Failed Res: " + JSON.stringify(res))
-        throw res
-    }
-    return(
-    
-            <GoogleLogin
-                // clientId={clientId}
-                // buttonText="Login"
-                // useOneTap={true}
-                onSuccess={onSuccess}
-                onError={() => onError}
-                // cookiePolicy={'single_host_origin'}
-                // isSignedIn={true}
-            />
-    
-    )
-}
+            <Modal
+                open={open}
+                onClose={handleClose}
+                aria-labelledby="modal-modal-title"
+                aria-describedby="modal-modal-description"
+            >
+                <Box sx={style}>
+                    <LoginPage closeModal={handleClose} />
+                </Box>
+            </Modal>
+        </>
+    );
+};
+
+export default LoginButton;
